+++ R/itemfit.R
@@ -326,8 +326,10 @@ itemfit <- function(x, fit_stats = 'S_X2', which.items = 1:extract.mirt(x, 'nite
         stopifnot(nrow(org) == length(which.items))
         model <- extract.mirt(mod, 'model')
         sv <- mod2values(mod)
-        retQ1 <- mySapply(1L:boot, pb_fun, mod=mod, N=N, sv=sv, itemtype=itemtype,
-                          which.items=which.items, draws=draws, model=model, ...)
+        retQ1 <- mySapply(1L:boot, pb_fun, progress=verbose,
+                          mod=mod, N=N, sv=sv, itemtype=itemtype,
+                          which.items=which.items, draws=draws, model=model,
+                          p.adjust=p.adjust, ...)
         if(nrow(retQ1) == 1L) retQ1 <- t(retQ1)
         Q1 <- (1 + rowSums(org$p.PV_Q1 > t(retQ1), na.rm = TRUE)) / (1 + boot)
         ret <- data.frame("p.PV_Q1_star"=p.adjust(Q1, method=p.adjust))
@@ -387,7 +389,8 @@ itemfit <- function(x, fit_stats = 'S_X2', which.items = 1:extract.mirt(x, 'nite
         stopifnot(length(org) == length(which.items))
         sv <- mod2values(mod)
         model <- extract.mirt(mod, 'model')
-        X2bs <- mySapply(1L:boot, pb_fun, mod=mod, N=N, model=model, is_NA=is_NA,
+        X2bs <- mySapply(1L:boot, pb_fun, progress=verbose,
+                         mod=mod, N=N, model=model, is_NA=is_NA,
                          itemtype=itemtype, sv=sv, which.items=which.items,
                          ETrange=ETrange, ETpoints=ETpoints, ...)
         if(nrow(X2bs) == 1L) X2bs <- t(X2bs)
@@ -461,7 +464,7 @@ itemfit <- function(x, fit_stats = 'S_X2', which.items = 1:extract.mirt(x, 'nite
         ret <- vector('list', x@Data$ngroups)
         if(is.null(Theta))
             Theta <- fscores(x, method=method, full.scores=TRUE, plausible.draws=impute,
-                             leave_missing=TRUE, ...)
+                             rotate = 'none', leave_missing=TRUE, ...)
         for(g in seq_len(x@Data$ngroups)){
             if(impute > 0L){
                 tmpTheta <- vector('list', impute)
@@ -504,7 +507,7 @@ itemfit <- function(x, fit_stats = 'S_X2', which.items = 1:extract.mirt(x, 'nite
     if(Zh || infit){
         if(is.null(Theta))
             Theta <- fscores(x, verbose=FALSE, full.scores=TRUE, method=method,
-                             leave_missing=TRUE, ...)
+                             leave_missing=TRUE, rotate = 'none', ...)
         prodlist <- attr(pars, 'prodlist')
         nfact <- x@Model$nfact + length(prodlist)
         fulldata <- x@Data$fulldata[[1L]]
@@ -557,7 +560,7 @@ itemfit <- function(x, fit_stats = 'S_X2', which.items = 1:extract.mirt(x, 'nite
     if(( (X2 || G2) || !is.null(empirical.plot) || !is.null(empirical.table)) && x@Model$nfact == 1L){
         if(is.null(Theta))
             Theta <- fscores(x, verbose=FALSE, full.scores=TRUE, method=method,
-                             leave_missing=TRUE, ...)
+                             rotate = 'none', leave_missing=TRUE, ...)
         nfact <- ncol(Theta)
         prodlist <- attr(pars, 'prodlist')
         fulldata <- x@Data$fulldata[[1L]]
@@ -790,7 +793,7 @@ itemfit <- function(x, fit_stats = 'S_X2', which.items = 1:extract.mirt(x, 'nite
                              y=c(as.numeric(Osub), as.numeric(Esub)),
                              type = rep(c('observed', 'expected'), each = prod(dim(Osub))),
                              category=rep(paste0('cat', 1L:ncol(Osub)), each=nrow(Osub)))
-            return(xyplot(y~Scores|category, df, type='l', group=df$type,
+            return(xyplot(y~Scores|category, df, type='b', group=df$type,
                    xlab = ifelse(S_X2.plot_raw.score, expression(Sum-Score), expression(theta)),
                    ylab=expression(P(theta)), auto.key=auto.key, par.strip.text=par.strip.text,
                    main = paste0("Observed vs Expected Values for Item ", S_X2.plot),