+++ R/mixedmirt.R
@@ -47,8 +47,8 @@
 #'   then the observations from \code{covdata} and \code{data} will be removed row-wise
 #'   via the \code{rowSums(is.na(covdata)) > 0}
 #' @param model an object returned from, or a string to be passed to, \code{mirt.model()}
-#'   to declare how the IRT model is to be estimated. See \code{\link{mirt.model}} for
-#'   more details
+#'   to declare how the IRT model is to be estimated. See \code{\link{mirt.model}} and
+#'   \code{\link{mirt}} for more detail
 #' @param fixed a right sided R formula for specifying the fixed effect (aka 'explanatory')
 #'   predictors from \code{covdata} and \code{itemdesign}. To estimate the intercepts for
 #'   each item the keyword \code{items} is reserved and automatically added to the \code{itemdesign}
@@ -129,6 +129,9 @@
 #' covdata <- data.frame(group, pseudoIQ)
 #' #use parallel computing
 #' mirtCluster()
+#'
+#' itemstats(data)
+#'
 #'
 #' #specify IRT model
 #' model <- 'Theta = 1-10'
@@ -266,6 +269,7 @@
 #' Theta <- matrix(c(rnorm(n, 0), rnorm(n, 1), rnorm(n, 2)))
 #' covdata <- data.frame(group=rep(c('g1','g2','g3'), each=n))
 #' dat <- simdata(a, d, N=n*3, Theta=Theta, itemtype = '2PL')
+#' itemstats(dat)
 #'
 #' #had we known the latent abilities, we could have computed the regression coefs
 #' summary(lm(Theta ~ covdata$group))
@@ -316,6 +320,7 @@
 #' group = factor(rep(paste0('G',1:cluster), each = N/cluster))
 #' covdata <- data.frame(group)
 #' dat <- simdata(a,d,N, itemtype = rep('2PL',10), Theta=matrix(Theta))
+#' itemstats(dat)
 #'
 #' # null model
 #' mod1 <- mixedmirt(dat, covdata, 1, fixed = ~ 0 + items, random = ~ 1|group)
@@ -344,7 +349,7 @@
 #' head(cbind(randef(mod3)$group, random_intercept))
 #'
 #' }
-mixedmirt <- function(data, covdata = NULL, model, fixed = ~ 1, random = NULL, itemtype = 'Rasch',
+mixedmirt <- function(data, covdata = NULL, model = 1, fixed = ~ 1, random = NULL, itemtype = 'Rasch',
                       lr.fixed = ~ 1, lr.random = NULL, itemdesign = NULL, constrain = NULL,
                       pars = NULL, return.design = FALSE, SE = TRUE, internal_constraints = TRUE,
                       technical = list(SEtol = 1e-4), ...)
@@ -434,7 +439,7 @@ mixedmirt <- function(data, covdata = NULL, model, fixed = ~ 1, random = NULL, i
                                 itemdesign=itemdesign, N=nrow(covdata))
     } else mr <- list()
     mixed.design <- list(fixed=mm, random=mr)
-    if(class(lr.random) == 'formula') lr.random <- list(lr.random)
+    if(is(lr.random, 'formula')) lr.random <- list(lr.random)
     if(length(lr.random) > 0L){
         lr.random <- make.randomdesign(random=lr.random, longdata=covdata,
                                        covnames=colnames(covdata), itemdesign=NULL,